+++ app/api/contact/route.js
@@ -1,0 +1,6 @@
+import nodemailer from 'nodemailer';
+
+
+export default async function handler(req,res){
+    
+}
